FEATURE: autotable: Allow hooking when table data is ready
This patch introduces prop :on-data-ready-sync used to pass an array
of callback functions, allowing to read or post-process the table
data once it is ready and before rendering.

Callbacks are meant NOT to be async, there is no await, hence the
name :on-data-ready-sync for the prop. Technically, if you know what
you do, it is okay to pass an async callback but the "loading"
progress animation won't await it.