FEATURE: Create a config dialog form
If the data configuration (mainly the data source url) is not found in the current route's query params, the user is prompted with a form dialog to fill for the config info. The app then redirects to apply the new config to the URL.

The handling of the config from the route (URL) was duplicated in a created() hook AND in the route watcher. The created() hook was rendered useless by using the "immediate" property of the route watcher, and was removed.