FEATURE: Create navigatable data tree & AutoObject components
In order to test the navigation through a multiple levels JSON file with arrays and objects, a mock file was added. In the config dialog, just use '/static/json-mock1.json' as source URL to use it.

For some utility methods of general use, a "utils" plugin was created. Was then added to this plugin : a method that creates a navigation link to an updated path in the API data, in order to navigate through the API data by clicking on this kind of links.

This component will be used alongside the AutoTable to switch between displays depending on the type of data at the current dataPath : AutoTable for an array and AutoObject for an object view.

The data displayed in either the AutoTable component, or the AutoObject component needed a rework in order to change its format. Indeed, depending on whether it contains sub levels of data (arrays or objects) or not, the data values needed to be changed in order to create HTML links to navigate to these sub levels of data. These links update the path and thus allow the navigation through all levels of depth in the global API data.

Alongside the tableData property, an objectData property was created with getters & setters, for future implementation regarding the AutoObject component.

Currently, any change in the layout or in the size of the window "breaks" the view because the height of the AutoTable and the AutoObject don't update. An event listener was added to both components to automatically re-compute their respective height on any resize or update to the component.

A breadcrumbs component was created to allow the user to see where the data in the view is located in the global data object. Every element on the current path is clickable and allows navigation back to previous levels in the data.

With the goal of having a collapsible side-panel with a "tree view" of the global data object, that can be navigated, a side-panel component name "TreePanel" with created alongside a button in the top nav to toggle its visibility.

A jsconfig file was created to enable intellisense to work with imports using the "@" identifier.

In order to generate unique IDs for the items in the Treeview component, the "vue-uuid" npm package was added to the project. Remember to run "npm install" or "yarn".

A tooltip was added to the tree-view toggle button in the breadcrumbs component.

In the table, single items could not directly be navigated to. In "generic" mode, an item index header name "#" was added, along with its formatting that simply creates a link to update the path with the index of the item.