FEATURE: AutoTable: add column order config
In the Column Configuration menu of the table component, a new option was added : it allows a column to be moved up or down in the current order, allowing for a reordering of the table headers. Once a column is moved, the new order is saved in local storage, and loaded on startup.