FEATURE: global: dissociate data management from AutoTable
The AutoTable component is meant to be an automated data table component, that should only display some given data with some functionalities.
A Data Manager plugin was created in order to handle the fetching & formatting of data from any API url, generating table headers etc, that was previously done by the AutoTable component.

The apiConfig object that contains the information necessary to fetch data from an API was made global and moved to the main.js file, instead of the App.vue.

The data-type prop is obsolete and needed to be removed