BUG/MINOR: autotable: Fix ignored column definition label
The column definition label, if defined, should be used as header
text. It was ignored, always falling back to the auto "ucfirst"
string of the column id.

The patch also removes the specific test that was made on "flag"
to produce an empty header text on that column. It requires to
explicitly define an empty string label for the column definition
in ViewInventory.