MINOR: autotable: Remove elevation to produce a flatter effect
The elevation-0 class on v-data-table had no effect, its removal do
not change anything. However, the patch removes the elevation on the
v-card container, to produce a flatter effect. It should be up to
parent component to decide if it wants an elevation of the table,
this is now possible.