FEATURE: autotable: Support cell render slot
I think this should be the Vue way to customize table cells instead
of using format(). It is far less convenient to use because the
column definition is separated from the template, this is not the
simplest to read... However, this is easiest way to generate DOM
when we want to include Vue components, events like @click, router
links, etc.