MINOR: autotable: Add bottom border on the last row
In the default style of the table (vuetify v-data-table), cells have
no bottom border and the footer area has a top border.

In the way AutoTable is mostly used, the tbody area expands to the
remaining height of the container and the footer area sticks to the
bottom. Due to the v-data-table style, the last body row has no bottom
border, giving the impression that it is not "closed". The patch
restores the bottom border on the last row.