BUG/MAJOR: AutoTable: improve table index links
In the AutoTable view, the index column is there to navigate through the array found at the given path.
If the root level is already an array, there's no preset path, and the formatManager did not take that into account when generating links for index entries. It now checks if there's an existing path before setting the target url.