FEATURE: AutoTable: display first elements of arrays/objects in a cell
Previously, when a table cell contained an array or an object, it would be formatted to show either "Array" or "Object" with a link to open the corresponding sub-level of the data.
This feature aims to improve UI and readability by showing the first element(s) of the array/object as a String to give the user a preview of the content instead.