BUG/MEDIUM: Restore table index column
Since the AutoTable update, the table headers are sefl-generated based on the data. This means that the generated headers will only feature properties that already exist on the data received from the API.
In order to restore the "index" column for ergonomy and (later) navigation, we used the config property named "customHeadersComputation".
It takes a callback with the headers as an argument, and can manipulate the array of headers, so an index header was added and placed at the start of the array in order to be the first column in the table.