BUG/MAJOR: Global: restore table navigation
In-table navigation was broken since the last AutoTable update. This restored this feature and improved it.
A full rework of the "generateHeaders" (now called "generateColumnDefinitions") was made. It now uses the new data structure for table items in the AutoTable, and the cell content navigation was updated to work with this structure.