BUG/MAJOR: Restore custom config URL handling
This feature was missing/broken since the massive AutoTable update and has now been restored. The naming was changed from "headers url" to "custom config url" as it now fetches the whole table config from a distant url and loads it into the AutoTable.
Some deprecated code was also removed as it's either no longer used, or redundant with some of the AutoTable's new features.