FEATURE/MAJOR: Upload a JSON file as a data source
This commit allows the user to upload a local JSON file as a data source, as opposed to a remote API source like before. 
This means the ConfigDialog and DataManager had to be (at least partially) reworked in order to welcome this change. The main App.vue file also needed an update to handle errors and compatibility