FEATURE: Add distant headers config support
This feature allows the user to specify a distant API URL to fetch a configuration JSON file for the headers of the table. It allows the user to specify which header will be shown, how its values will be formatted, and what CSS to apply to its values.
Note: Specifying a data-type is mandatory in order to have apply some CSS to headers and their values, since CSS loading is related to data-type.

Headers were given an internal unique key "hid" to identify them across the app, for unicity purposes in the Autotable an

Created a second tab in the configuration dialog for Column configuration. For now, this consists in a multi-select dropdown to choose which table columns to display or not, given the current data. The selection is then applied to the Autotable and saved in Local Storage, to be restored on any page load on the same environment.