MEDIUM: AutoTableMenu: Rework menu structure & style make entries scrollable
In order to prevent the config menu from overflowing past the AutoTable's maximum height, the menu was made scrollable, its structure was changed so that the menu Header and Footer are now fixed while only the entry list is being scrolled.
The style was also reworked to accomodate for the changes in position and to improve readability.