MEDIUM: Restore AutoTable's config "api" and "search" and handle prop overrides
The "api" and "search" properties of the AutoTable's config object were previously replaced by props in order to separate them from the config.
This commit restores the config.api and config.search properties, while still keeping the props as an alternative way to configure the AutoTable.
The CONFIG API has PRIORITY over the api prop.
In the case of an api override (both properties are provided but the config api is used), the AutoTable now fires an event, alerting its parent component that there was an override.
The parent component (Main here) now handles the override by updating the DataManager's config and redirecting to update the url query params.