MINOR: autotable: Keep it simple, don't use loops when there is no need
Array.includes() is a loop and is not natural to read.

The patch does not touch the logic but those typeof tests should be
reviewed. Why not tests for object, maybe arrays? Is that still required
since there is kind of the same code in the demo viewer interface when
generating columns definition (dataManager.js)?