MINOR: autotable: Prevent text selection on range select
The commit adds an internal mousedown event handler on body rows.
It is only used to prevent user-select when selecting multiple rows,
ie. when holding the shift key. It's an alternative to using CSS
"user-select: none" with benefit that user manual text selection is
still possible without hassle. It was tested successfully on Chrome
and Firefox. It has been reported it does not work with legacy IE,
which is fine.